% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeTB_CE_matrix_by_age.R
\name{costeff_matrix}
\alias{costeff_matrix}
\title{costeff_matrix}
\usage{
costeff_matrix(
  start_age = 34,
  n_pop = 1000,
  n_LTBI = 105.739,
  QALY0 = 924.77849,
  cost0 = 17051.77
)
}
\arguments{
\item{start_age}{single value starting age in years}

\item{n_pop}{initial total population size}

\item{n_LTBI}{initial number of LTBI cases}

\item{QALY0}{startup QALY}

\item{cost0}{startup cost}
}
\value{
dataframe
}
\description{
This replicates the population active TB calculation
in Manabu's Excel model.
}
