% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ce_tree_long_df.R
\name{create_ce_tree_long_df}
\alias{create_ce_tree_long_df}
\title{create_ce_tree_long_df}
\usage{
create_ce_tree_long_df(
  tree_list = NA,
  tree_mat = NA,
  label_probs,
  label_costs,
  pname_from_to,
  cname_from_to
)
}
\arguments{
\item{cname_from_to}{}
}
\value{

}
\description{
Long input data for define_model().
}
\examples{

}
