% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ltbi_heemod.R
\name{create_ltbi_heemod}
\alias{create_ltbi_heemod}
\title{create_ltbi_heemod}
\usage{
create_ltbi_heemod(
  age_init = 34,
  pReact_comp = 0.0006779,
  pReact_incomp = 0.0015301,
  pReact = 0.0019369,
  TB_cost = 4925.76,
  d = 0.035,
  N = 1
)
}
\arguments{
\item{age_init}{Starting age}

\item{pReact_comp}{TB after completed LTBI treatment}

\item{pReact_incomp}{TB after LTBI treatment drop-out}

\item{TB_cost}{Cost of TB treatment (£)}

\item{d}{Discount factor}

\item{N}{Number of runs}

\item{pReactB}{TB after no treatment}
}
\value{
Returns a function with initial sub-populations as input.
}
\description{
Create a heemod Markov model for the LTBI test population.
Transitions happen at the beginning of each year (equivalent to transition happening at
the end + ignoring the first year) with method = "beginning".
Since with this method the first year is actually the second,
costs should be discounted from the start with the argument first = TRUE in discount().
See: https://cran.r-project.org/web/packages/heemod/vignettes/d_non_homogeneous.html
}
\examples{

create_ltbi_heemod()

}
