% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pop_heemod_PSA.R
\name{heemod_init_pop_PSA}
\alias{heemod_init_pop_PSA}
\title{Probability sensitivity analysis for initial population}
\usage{
heemod_init_pop_PSA(heemod_model, init_states)
}
\arguments{
\item{heemod_model}{function to run heemod with parameters set}

\item{init_states}{data.frame of scenario starting state population
scenario as rows}
}
\description{
heemod Markov model taking decision tree sub-populations
as input.
Make sure inits in correct order.
}
